requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT 支持用于前后端完全分离的无状态认证
djangorestframework-simplejwt==5.3.1

# API 文档
drf-spectacular==0.27.0

# Excel 导出
openpyxl==3.1.2

# 生产部署
gunicorn==21.2.0
whitenoise==6.6.0

# 系统监控
psutil==5.9.8

# HTTP 请求（节假日 API）
requests==2.31.0
